refactor(push): separate notification navigation
